fix system_temp_directory
Patch proposed to llvm project:

In lib/Support/Path.h the documentation for system_temp_directory contains:
@param erasedOnReboot […] This parameter will be ignored if the user or system
has set the typical environment variable […]

In my eyes on Unix the method currently does the exact opposite: It behaves the
other way around by ignoring the environment variable if the parameter is set.

I included a patch that changes this.
It does what I personally need, but I’m not 100% sure yet that this is
generally the desired behavior.

If it is not, I feel that instead of this patch the documentation in Path.h
should be fixed or clarified.

I’m happy to include test coverage for this if needed, but I would first like
to confirm the desired behavior.